Rename tunnel, admin, and console gems.
