+++ admin/Rakefile
@@ -1,7 +1,7 @@
 require "rake"
 
 $LOAD_PATH.unshift File.expand_path("../lib", __FILE__)
-require "admin-vmc-plugin/version"
+require "appfog-admin-vmc-plugin/version"
 
 task :default => :spec
 
@@ -12,21 +12,21 @@ desc "Run integration tests"
 task :test => ["bundler:install", "test:integration"]
 
 task :build do
-  sh "gem build admin-vmc-plugin.gemspec"
+  sh "gem build appfog-admin-vmc-plugin.gemspec"
 end
 
 task :install => :build do
-  sh "gem install --local admin-vmc-plugin-#{VMCAdmin::VERSION}.gem"
+  sh "gem install --local appfog-admin-vmc-plugin-#{VMCAdmin::VERSION}.gem"
 end
 
 task :uninstall do
-  sh "gem uninstall admin-vmc-plugin"
+  sh "gem uninstall appfog-admin-vmc-plugin"
 end
 
 task :reinstall => [:uninstall, :install]
 
 task :release => :build do
-  sh "gem push admin-vmc-plugin-#{VMCAdmin::VERSION}.gem"
+  sh "gem push appfog-admin-vmc-plugin-#{VMCAdmin::VERSION}.gem"
 end
 
 namespace "bundler" do
