--- admin/lib/admin-vmc-plugin/plugin.rb
@@ -1,2 +1,0 @@
-require "admin-vmc-plugin/curl"
-require "admin-vmc-plugin/service_auth_token"
