+++ admin/lib/appfog-admin-vmc-plugin/curl.rb
