+++ admin/lib/appfog-admin-vmc-plugin/plugin.rb
@@ -1,0 +1,2 @@
+require "appfog-admin-vmc-plugin/curl"
+require "appfog-admin-vmc-plugin/service_auth_token"
