+++ admin/lib/appfog-admin-vmc-plugin/service_auth_token.rb
