+++ admin/lib/appfog-admin-vmc-plugin/version.rb
