+++ console/lib/appfog-console-vmc-plugin/version.rb
