+++ tunnel/Rakefile
@@ -1,7 +1,7 @@
 require "rake"
 
 $LOAD_PATH.unshift File.expand_path("../lib", __FILE__)
-require "tunnel-vmc-plugin/version"
+require "appfog-tunnel-vmc-plugin/version"
 
 task :default => :spec
 
@@ -12,21 +12,21 @@ desc "Run integration tests"
 task :test => ["bundler:install", "test:integration"]
 
 task :build do
-  sh "gem build tunnel-vmc-plugin.gemspec"
+  sh "gem build appfog-tunnel-vmc-plugin.gemspec"
 end
 
 task :install => :build do
-  sh "gem install --local tunnel-vmc-plugin-#{VMCTunnel::VERSION}.gem"
+  sh "gem install --local appfog-tunnel-vmc-plugin-#{VMCTunnel::VERSION}.gem"
 end
 
 task :uninstall do
-  sh "gem uninstall tunnel-vmc-plugin"
+  sh "gem uninstall appfog-tunnel-vmc-plugin"
 end
 
 task :reinstall => [:uninstall, :install]
 
 task :release => :build do
-  sh "gem push tunnel-vmc-plugin-#{VMCTunnel::VERSION}.gem"
+  sh "gem push appfog-tunnel-vmc-plugin-#{VMCTunnel::VERSION}.gem"
 end
 
 namespace "bundler" do
