--- tunnel/lib/tunnel-vmc-plugin/version.rb
@@ -1,3 +1,0 @@
-module VMCTunnel
-  VERSION = "0.2.0.1".freeze
-end
